feat: add layout resize
