+++ src/components/ListCovid/ListCovid.tsx
@@ -2,6 +2,12 @@ import React from 'react';
 
 import styles from '@/components/ListCovid/ListCovid.scss';
 
-const ListCovid = (): JSX.Element => <div className={styles['list-covid']}>ListCovid</div>;
+import ComponentLayout from '../layout';
+
+const ListCovid = (): JSX.Element => (
+  <ComponentLayout>
+    <div className={styles['list-covid']}>ListCovid</div>
+  </ComponentLayout>
+);
 
 export default ListCovid;
