+++ src/components/TableCovid/TableCovid.tsx
@@ -2,6 +2,12 @@ import React from 'react';
 
 import styles from '@/components/TableCovid/TableCovid.scss';
 
-const TableCovid = (): JSX.Element => <div className={styles['table-covid']}>TableCovid</div>;
+import ComponentLayout from '../layout';
+
+const TableCovid = (): JSX.Element => (
+  <ComponentLayout>
+    <div className={styles['table-covid']}>TableCovid</div>
+  </ComponentLayout>
+);
 
 export default TableCovid;
