+++ src/components/layout/ComponentLayout.scss
@@ -1,0 +1,36 @@
+.covid {
+  position: relative;
+  width: 100%;
+  height: 100%;
+  background-color: green;
+}
+
+.open-covid {
+  position: absolute;
+  z-index: 3;
+  top: 0.5%;
+  left: 0.5%;
+  width: 99%;
+  height: 99%;
+  background-color: green;
+}
+
+.close {
+  position: absolute;
+  top: 1%;
+  right: 1%;
+
+  &:hover {
+    cursor: pointer;
+  }
+}
+
+.expand-li-icon {
+  display: block;
+  width: 1rem;
+  height: 1rem;
+
+  &:hover {
+    fill: rgba(17, 17, 17, 0.651);
+  }
+}
