+++ src/components/layout/index.tsx
@@ -1,0 +1,3 @@
+import ComponentLayout from '@/components/layout/ComponentLayout';
+
+export default ComponentLayout;
